Trait multicollateral_bonding_curve_pool::pallet::Config

pub trait Config: Config + Config + Config + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type LiquidityProxy: LiquidityProxyTrait<Self::DEXId, Self::AccountId, AssetIdOf<Self>>;
    type EnsureDEXManager: EnsureDEXManager<Self::DEXId, Self::AccountId, DispatchError>;
    type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, AssetIdOf<Self>, DispatchError>;
    type PriceToolsPallet: PriceToolsProvider<AssetIdOf<Self>>;
    type VestedRewardsPallet: Vesting<Self::AccountId, AssetIdOf<Self>>;
    type TradingPairSourceManager: TradingPairSourceManager<Self::DEXId, AssetIdOf<Self>>;
    type BuyBackHandler: BuyBackHandler<Self::AccountId, AssetIdOf<Self>>;
    type GetBuyBackAssetId: Get<AssetIdOf<Self>>;
    type IrreducibleReserve: Get<Percent>;
    type WeightInfo: WeightInfo;
    type AssetInfoProvider: AssetInfoProvider<AssetIdOf<Self>, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>;

    const RETRY_DISTRIBUTION_FREQUENCY: BlockNumberFor<Self>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type LiquidityProxy: LiquidityProxyTrait<Self::DEXId, Self::AccountId, AssetIdOf<Self>>

type EnsureDEXManager: EnsureDEXManager<Self::DEXId, Self::AccountId, DispatchError>

type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, AssetIdOf<Self>, DispatchError>

type PriceToolsPallet: PriceToolsProvider<AssetIdOf<Self>>

type VestedRewardsPallet: Vesting<Self::AccountId, AssetIdOf<Self>>

type TradingPairSourceManager: TradingPairSourceManager<Self::DEXId, AssetIdOf<Self>>

type BuyBackHandler: BuyBackHandler<Self::AccountId, AssetIdOf<Self>>

type GetBuyBackAssetId: Get<AssetIdOf<Self>>

type IrreducibleReserve: Get<Percent>

Percent of reserve which is not involved in swap

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type AssetInfoProvider: AssetInfoProvider<AssetIdOf<Self>, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>

To retrieve asset info

Required Associated Constants

const RETRY_DISTRIBUTION_FREQUENCY: BlockNumberFor<Self>